page changes for notifications
