+++ server/routes/notification.js
@@ -8,32 +8,6 @@ const linkSubToParent = require('../data/linkToParent');
 const ownershipChangeRequests = require('../data/ownership');
 const notifications = require('../data/notifications');
 
-/**
- * Method will fetch the notification details.
- * @param notification
- */
-const getNotificationTypes = async (req, res) => {
-  const notificationTypes = await notifications.getNotificationTypes();
-  return res.status(200).send(notificationTypes);
-};
-
-const createNotificationType = async (req, res) => {
-  try {
-    if (req.params === null) {
-      return res.status(400).send({ message: 'No new type was provided' });
-    }
-
-    await notifications.createNotificationType(req.body);
-
-    const newType = notifications.selectNotificationTypeByTypeName(req.body.type);
-    return res.status(200).send(newType);
-  } catch (e) {
-    return res.status(500).send({
-      message: e.message,
-    });
-  }
-};
-
 const getEstablishmentNotifications = async (req, res) => {
   const establishmentNotifications = await notifications.selectNotificationByEstablishment(req.params.establishmentUid);
   return res.status(200).send(establishmentNotifications);
@@ -252,8 +226,6 @@ const sendUserNotification = async (req, res) => {
   }
 };
 
-router.route('/type').get(getNotificationTypes);
-router.route('/type').post(createNotificationType);
 router.route('/user/:userUid').post(sendUserNotification);
 router.route('/establishment/:establishmentUid').get(getEstablishmentNotifications);
 router.route('/establishment/:establishmentUid').post(sendEstablishmentNotification);
