+++ src/app/features/notifications/notification-cancelled/notification-cancelled.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NotificationsService } from '@core/services/notifications/notifications.service';
 import { Subscription } from 'rxjs';
@@ -7,30 +7,18 @@ import { Subscription } from 'rxjs';
   selector: 'app-notification-cancelled',
   templateUrl: './notification-cancelled.component.html',
 })
-export class NotificationCancelledComponent implements OnInit, OnDestroy {
+export class NotificationCancelledComponent implements OnInit {
+  @Input() public notification;
   public cancelledMessage;
-  private subscriptions: Subscription = new Subscription();
-  constructor(
-    private route: ActivatedRoute,
-    private notificationsService: NotificationsService,
-    private router: Router
-  ) {}
+  constructor(private router: Router) {}
 
   ngOnInit() {
-    const notificationUid = this.route.snapshot.params.notificationuid;
-    this.subscriptions.add(
-      this.notificationsService.getNotificationDetails(notificationUid).subscribe(details => {
-        if (details.type === 'LINKTOPARENTREQUEST') {
-          this.cancelledMessage = details.typeContent.requestorName + ' cancelled their request to link to you.';
-        } else {
-          this.cancelledMessage =
-            details.typeContent.requestorName + ' cancelled the request to change ownership of the data.';
-        }
-      })
-    );
-  }
-  ngOnDestroy(): void {
-    this.subscriptions.unsubscribe();
+    if (this.notification.type === 'LINKTOPARENTREQUEST') {
+      this.cancelledMessage = this.notification.typeContent.requestorName + ' cancelled their request to link to you.';
+    } else {
+      this.cancelledMessage =
+        this.notification.typeContent.requestorName + ' cancelled the request to change ownership of the data.';
+    }
   }
 
   public returnToHome() {
