+++ src/app/features/notifications/notification-link-to-parent/notification-link-to-parent.component.html
@@ -1,20 +1,3 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l">
-      <span class="govuk-caption-xl">{{ workplace?.name }}</span>
-      Notification
-    </h1>
-  </div>
-</div>
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds" *ngIf="notification">
-    <h2 class="govuk-heading-s govuk-util__inline-block govuk-!-margin-right-2">From:</h2>
-    <p class="govuk-util__inline-block">{{ requestorName }}, {{ notification.created | longDate }}</p>
-    <br />
-    <h2 class="govuk-heading-s govuk-util__inline-block govuk-!-margin-right-2">Subject:</h2>
-    <p class="govuk-util__inline-block">{{ notification?.type | notificationType }}</p>
-  </div>
-</div>
 <ng-container *ngIf="notification">
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
@@ -92,20 +75,6 @@
           </tr>
         </tbody>
       </table>
-    </div>
-  </div>
-  <div class="govuk-grid-row" *ngIf="displayActionButtons">
-    <div class="govuk-grid-column-two-thirds">
-      <div class="govuk-!-margin-top-2">
-        <button (click)="approveRequest()" type="submit" class="govuk-button">Approve request</button>
-        <button
-          (click)="rejectRequest($event)"
-          type="button"
-          class="govuk-button govuk-button--secondary govuk-util__float-right"
-        >
-          Reject request
-        </button>
-      </div>
     </div>
   </div>
 </ng-container>
