+++ src/app/features/notifications/notification-list/notification-list.component.html
@@ -21,7 +21,10 @@
             <span *ngIf="!notification.isViewed" class="notification-list-item__icon">
               <span class="govuk-visually-hidden">Unread: </span></span
             >
-            <a
+            <a [routerLink]="['/notifications', notification.notificationUid]" class="notification-list-item__link">{{
+              notification.type | notificationType
+            }}</a>
+            <!-- <a
               *ngIf="notification.type === 'OWNERCHANGE'"
               [routerLink]="[notification.notificationUid]"
               class="notification-list-item__link"
@@ -48,7 +51,7 @@
               [routerLink]="['/notifications', 'notification-become-a-parent', notification.notificationUid]"
               class="notification-list-item__link"
             >{{ notification.type | notificationType }}</a
-            >
+            > -->
           </span>
         </td>
         <td class="govuk-table__cell">{{ notification.created | longDate }}</td>
