+++ src/app/features/notifications/notification/notification.component.html
@@ -10,7 +10,8 @@
   <div class="govuk-grid-column-two-thirds" *ngIf="notification">
     <h2 class="govuk-heading-s govuk-util__inline-block govuk-!-margin-right-2">From:</h2>
     <p class="govuk-util__inline-block">
-      {{ notification.typeContent.requestorName }}, {{ notification.created | longDate }}
+      {{ notification.typeContent.requestorName ? notification.typeContent.requestorName : 'Skills For Care' }},
+      {{ notification.created | longDate }}
     </p>
     <br />
     <h2 class="govuk-heading-s govuk-util__inline-block govuk-!-margin-right-2">Subject:</h2>
@@ -18,113 +19,32 @@
   </div>
 </div>
 <ng-container *ngIf="notification">
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-      <div [ngSwitch]="notification.typeContent.approvalStatus">
-        <div *ngSwitchCase="'REQUESTED'">
-          <p>You have a request to transfer ownership of data to {{ ownerShipRequestedTo }}.</p>
-          <p>Approve the request to set the following permissions:</p>
-        </div>
-        <div *ngSwitchCase="'CANCELLED'">
-          <p>You have a request to transfer ownership of data to {{ ownerShipRequestedTo }}.</p>
-          <p>Approve the request to set the following permissions:</p>
-        </div>
-
-        <div *ngSwitchCase="'DENIED'">
-          <ng-container *ngIf="isWorkPlaceIsRequester; else nonWorkPlaceRequester">
-            <p>You rejected the request to transfer ownership of data to {{ ownerShipRequestedTo }}</p>
-            <h2 class="govuk-heading-s">Reason</h2>
-            <ng-container *ngIf="notification.typeContent.rejectionReason; else noReason">
-              <p>{{ notification.typeContent.rejectionReason }}</p>
-            </ng-container>
-            <ng-template #noReason>
-              <p>No reason provided.</p>
-            </ng-template>
-          </ng-container>
-          <ng-template #nonWorkPlaceRequester>
-            <p>{{ ownerShipRequestedFrom }} has rejected your request to become the data owner.</p>
-            <h2 class="govuk-heading-s">Reason</h2>
-            <ng-container *ngIf="notification.typeContent.rejectionReason; else noReason">
-              <p>{{ notification.typeContent.rejectionReason }}</p>
-            </ng-container>
-            <ng-template #noReason>
-              <p>No reason provided.</p>
-            </ng-template>
-          </ng-template>
-        </div>
-        <div *ngSwitchCase="'APPROVED'">
-          <ng-container *ngIf="isWorkPlaceIsRequester; else nonWorkPlaceRequester">
-            <p>You approved the request to transfer ownership of data to {{ ownerShipRequestedFrom }}</p>
-          </ng-container>
-          <ng-template #nonWorkPlaceRequester>
-            <p>
-              {{ ownerShipRequestedTo }} has approved your request to become the data owner. You're now free to edit
-              workplace and staff data.
-            </p>
-          </ng-template>
-        </div>
-        <div *ngSwitchDefault></div>
-      </div>
-
-      <table class="govuk-table govuk-table__with-action" *ngIf="notification.typeContent.approvalStatus !== 'DENIED'">
-        <thead class="govuk-table__head">
-          <tr class="govuk-table__row">
-            <th scope="col" class="govuk-table__header">Workplace</th>
-            <th scope="col" class="govuk-table__header">Permissions</th>
-          </tr>
-        </thead>
-        <tbody class="govuk-tab">
-          <ng-container *ngIf="notification.typeContent.approvalStatus !== 'APPROVED'; else approved">
-            <tr class="govuk-table__row govuk-error-table__row--no-border">
-              <td class="govuk-table__cell">
-                {{ ownerShipRequestedFrom }}
-              </td>
-
-              <ng-container *ngIf="notification.typeContent.permissionRequest === 'None'; else notNone">
-                <td class="govuk-table__cell">
-                  {{ notification.typeContent.permissionRequest | dataViewPermissions }}
-                </td>
-              </ng-container>
-              <ng-template #notNone>
-                <td class="govuk-table__cell">
-                  View {{ notification.typeContent.permissionRequest | dataViewPermissions | lowercase }}
-                </td>
-              </ng-template>
-            </tr>
-            <tr class="govuk-table__row govuk-error-table__row--no-border" *ngIf="notification">
-              <td class="govuk-table__cell">
-                {{ ownerShipRequestedTo }}
-              </td>
-              <td class="govuk-table__cell">Edit workplace and staff records</td>
-            </tr>
-          </ng-container>
-          <ng-template #approved>
-            <tr class="govuk-table__row govuk-error-table__row--no-border">
-              <td class="govuk-table__cell">
-                {{ ownerShipRequestedTo }}
-              </td>
-
-              <ng-container *ngIf="notification.typeContent.permissionRequest === 'None'; else notNone">
-                <td class="govuk-table__cell">
-                  {{ notification.typeContent.permissionRequest | dataViewPermissions }}
-                </td>
-              </ng-container>
-              <ng-template #notNone>
-                <td class="govuk-table__cell">
-                  View {{ notification.typeContent.permissionRequest | dataViewPermissions | lowercase }}
-                </td>
-              </ng-template>
-            </tr>
-            <tr class="govuk-table__row govuk-error-table__row--no-border" *ngIf="notification">
-              <td class="govuk-table__cell">
-                {{ ownerShipRequestedFrom }}
-              </td>
-              <td class="govuk-table__cell">Edit workplace and staff records</td>
-            </tr>
-          </ng-template>
-        </tbody>
-      </table>
-    </div>
+  <div *ngIf="notification.type === 'OWNERCHANGE'">
+    <app-notification-owner-change
+      [notification]="notification"
+      [events]="eventsSubject.asObservable()"
+    ></app-notification-owner-change>
+  </div>
+  <div
+    *ngIf="
+      notification.type === 'LINKTOPARENTREQUEST' ||
+      notification.type === 'LINKTOPARENTAPPROVED' ||
+      notification.type === 'LINKTOPARENTREJECTED'
+    "
+  >
+    <app-notification-link-to-parent
+      [notification]="notification"
+      [events]="eventsSubject.asObservable()"
+    ></app-notification-link-to-parent>
+  </div>
+  <div *ngIf="notification.type === 'DELINKTOPARENT'">
+    <app-notification-delink-to-parent [notification]="notification"></app-notification-delink-to-parent>
+  </div>
+  <div *ngIf="notification.type === 'BECOMEAPARENT'">
+    <app-notification-become-a-parent
+      [notification]="notification"
+      [events]="eventsSubject.asObservable()"
+    ></app-notification-become-a-parent>
   </div>
   <div class="govuk-grid-row" *ngIf="displayActionButtons">
     <div class="govuk-grid-column-two-thirds">
