+++ src/app/features/notifications/notification/notification.component.ts
@@ -10,7 +10,7 @@ import { EstablishmentService } from '@core/services/establishment.service';
 import { NotificationsService } from '@core/services/notifications/notifications.service';
 import { PermissionsService } from '@core/services/permissions/permissions.service';
 import { RejectRequestDialogComponent } from '@shared/components/reject-request-dialog/reject-request-dialog.component';
-import { Subscription } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 
 const OWNERSHIP_APPROVED = 'OWNERCHANGEAPPROVED';
 const OWNERSHIP_REJECTED = 'OWNERCHANGEREJECTED';
@@ -31,6 +31,8 @@ export class NotificationComponent implements OnInit, OnDestroy {
   public ownerShipRequestedTo: string;
   public isSubWorkplace: boolean;
 
+  eventsSubject: Subject<string> = new Subject<string>();
+
   constructor(
     private route: ActivatedRoute,
     private breadcrumbService: BreadcrumbService,
@@ -49,7 +51,7 @@ export class NotificationComponent implements OnInit, OnDestroy {
 
     this.notificationsService.getNotificationDetails(this.notificationUid).subscribe((details) => {
       this.notification = details;
-
+      console.log(this.notification);
       this.isSubWorkplace =
         this.workplace.isParent && this.workplace.uid === this.establishmentService.primaryWorkplace.uid ? true : false;
 
@@ -72,57 +74,68 @@ export class NotificationComponent implements OnInit, OnDestroy {
     });
     this.setNotificationViewed(this.notificationUid);
   }
-  public approveRequest() {
-    if (this.notification) {
-      if (this.notification.typeContent.approvalStatus === 'CANCELLED') {
-        this.router.navigate(['/notifications/notification-cancelled', this.notification.notificationUid]);
-        return true;
-      }
 
-      const requestParameter = {
-        ownerRequestChangeUid: this.notification.typeContent.ownerChangeRequestUID,
-        approvalStatus: 'APPROVED',
-        approvalReason: '',
-        rejectionReason: null,
-        type: OWNERSHIP_APPROVED,
-        exsistingNotificationUid: this.notificationUid,
-        requestedOwnership: this.notification.typeContent.permissionRequest,
-      };
-      this.subscriptions.add(
-        this.notificationsService
-          .approveOwnership(this.notification.typeContent.ownerChangeRequestUID, requestParameter)
-          .subscribe(
-            (request) => {
-              if (request) {
-                this.establishmentService.getEstablishment(this.workplace.uid).subscribe((workplace) => {
-                  if (workplace) {
-                    this.permissionsService.getPermissions(this.workplace.uid).subscribe((hasPermission) => {
-                      if (hasPermission) {
-                        this.permissionsService.setPermissions(this.workplace.uid, hasPermission.permissions);
-                        this.establishmentService.setState(workplace);
-                        this.establishmentService.setPrimaryWorkplace(workplace);
-                        this.router.navigate(['/dashboard']);
-                        this.alertService.addAlert({
-                          type: 'success',
-                          message: `Your decision to transfer ownership of data has been sent to
-                      ${this.notification.typeContent.requestorName} `,
-                        });
-                      }
-                    });
-                  }
-                });
-                this.notificationsService.getAllNotifications(this.workplace.uid).subscribe((notify) => {
-                  this.notificationsService.notifications$.next(notify);
-                });
-              }
-            },
-            (error) => {
-              console.error(error.error.message);
-            },
-          ),
-      );
-    }
+  public approveRequest() {
+    this.eventsSubject.next('APPROVE');
   }
+
+  public rejectRequest($event: Event) {
+    $event.preventDefault();
+    this.eventsSubject.next('REJECT');
+  }
+
+  // public approveRequest() {
+  //   if (this.notification) {
+  //     if (this.notification.typeContent.approvalStatus === 'CANCELLED') {
+  //       this.router.navigate(['/notifications/notification-cancelled', this.notification.notificationUid]);
+  //       return true;
+  //     }
+
+  //     const requestParameter = {
+  //       ownerRequestChangeUid: this.notification.typeContent.ownerChangeRequestUID,
+  //       approvalStatus: 'APPROVED',
+  //       approvalReason: '',
+  //       rejectionReason: null,
+  //       type: OWNERSHIP_APPROVED,
+  //       exsistingNotificationUid: this.notificationUid,
+  //       requestedOwnership: this.notification.typeContent.permissionRequest,
+  //     };
+  //     this.subscriptions.add(
+  //       this.notificationsService
+  //         .approveOwnership(this.notification.typeContent.ownerChangeRequestUID, requestParameter)
+  //         .subscribe(
+  //           (request) => {
+  //             if (request) {
+  //               this.establishmentService.getEstablishment(this.workplace.uid).subscribe((workplace) => {
+  //                 if (workplace) {
+  //                   this.permissionsService.getPermissions(this.workplace.uid).subscribe((hasPermission) => {
+  //                     if (hasPermission) {
+  //                       this.permissionsService.setPermissions(this.workplace.uid, hasPermission.permissions);
+  //                       this.establishmentService.setState(workplace);
+  //                       this.establishmentService.setPrimaryWorkplace(workplace);
+  //                       this.router.navigate(['/dashboard']);
+  //                       this.alertService.addAlert({
+  //                         type: 'success',
+  //                         message: `Your decision to transfer ownership of data has been sent to
+  //                     ${this.notification.typeContent.requestorName} `,
+  //                       });
+  //                     }
+  //                   });
+  //                 }
+  //               });
+  //               this.notificationsService.getAllNotifications(this.workplace.uid).subscribe((notify) => {
+  //                 this.notificationsService.notifications$.next(notify);
+  //               });
+  //             }
+  //           },
+  //           (error) => {
+  //             console.error(error.error.message);
+  //           },
+  //         ),
+  //     );
+  //   }
+  // }
+
   private setNotificationViewed(notificationUid) {
     this.subscriptions.add(
       this.notificationsService.setNoticationViewed(notificationUid).subscribe(
@@ -141,53 +154,53 @@ export class NotificationComponent implements OnInit, OnDestroy {
     );
   }
 
-  public rejectRequest($event: Event) {
-    if (this.notification) {
-      if (this.notification.typeContent.approvalStatus === 'CANCELLED') {
-        this.router.navigate(['/notifications/notification-cancelled', this.notification.notificationUid]);
-        return true;
-      }
-      $event.preventDefault();
-      const dialog = this.dialogService.open(RejectRequestDialogComponent, this.notification);
-      dialog.afterClosed.subscribe((requestRejected) => {
-        if (requestRejected) {
-          this.rejectPermissionRequest(requestRejected);
-        }
-      });
-    }
-  }
+  // public rejectRequest() {
+  //   if (this.notification) {
+  //     if (this.notification.typeContent.approvalStatus === 'CANCELLED') {
+  //       this.router.navigate(['/notifications/notification-cancelled', this.notification.notificationUid]);
+  //       return true;
+  //     }
 
-  private rejectPermissionRequest(requestRejected) {
-    const requestParameter = {
-      ownerRequestChangeUid: this.notification.typeContent.ownerChangeRequestUID,
-      approvalStatus: 'DENIED',
-      rejectionReason: requestRejected.rejectionReason,
-      type: OWNERSHIP_REJECTED,
-      exsistingNotificationUid: this.notification.notificationUid,
-    };
-    this.subscriptions.add(
-      this.notificationsService
-        .approveOwnership(this.notification.typeContent.ownerChangeRequestUID, requestParameter)
-        .subscribe(
-          (request) => {
-            if (request) {
-              this.notificationsService.getAllNotifications(this.workplace.uid).subscribe((notify) => {
-                this.notificationsService.notifications$.next(notify);
-              });
-              this.router.navigate(['/dashboard']);
-              this.alertService.addAlert({
-                type: 'success',
-                message: `Your decision to transfer ownership of data has been sent to
-                  ${this.notification.typeContent.requestorName} `,
-              });
-            }
-          },
-          (error) => {
-            console.log('Could not update notification.');
-          },
-        ),
-    );
-  }
+  //     const dialog = this.dialogService.open(RejectRequestDialogComponent, this.notification);
+  //     dialog.afterClosed.subscribe((requestRejected) => {
+  //       if (requestRejected) {
+  //         this.rejectPermissionRequest(requestRejected);
+  //       }
+  //     });
+  //   }
+  // }
+
+  // private rejectPermissionRequest(requestRejected) {
+  //   const requestParameter = {
+  //     ownerRequestChangeUid: this.notification.typeContent.ownerChangeRequestUID,
+  //     approvalStatus: 'DENIED',
+  //     rejectionReason: requestRejected.rejectionReason,
+  //     type: OWNERSHIP_REJECTED,
+  //     exsistingNotificationUid: this.notification.notificationUid,
+  //   };
+  //   this.subscriptions.add(
+  //     this.notificationsService
+  //       .approveOwnership(this.notification.typeContent.ownerChangeRequestUID, requestParameter)
+  //       .subscribe(
+  //         (request) => {
+  //           if (request) {
+  //             this.notificationsService.getAllNotifications(this.workplace.uid).subscribe((notify) => {
+  //               this.notificationsService.notifications$.next(notify);
+  //             });
+  //             this.router.navigate(['/dashboard']);
+  //             this.alertService.addAlert({
+  //               type: 'success',
+  //               message: `Your decision to transfer ownership of data has been sent to
+  //                 ${this.notification.typeContent.requestorName} `,
+  //             });
+  //           }
+  //         },
+  //         (error) => {
+  //           console.log('Could not update notification.');
+  //         },
+  //       ),
+  //   );
+  // }
 
   ngOnDestroy(): void {
     this.subscriptions.unsubscribe();
